make rpc api compatible with React Native
