+++ libs/ledger-live-common/src/families/evm/api/rpc.ios.ts
@@ -1,0 +1,5 @@
+/** ⚠️ keep this order of import. @see https://docs.ethers.io/v5/cookbook/react-native/#cookbook-reactnative ⚠️ */
+import "react-native-get-random-values";
+import "@ethersproject/shims";
+export * from "./rpc.common";
+export * as default from "./rpc.common";
